requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
filelock==3.21.0
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numba==0.63.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
pycparser==3.0
regex==2026.1.15
requests==2.32.5
scipy==1.15.3
sounddevice==0.5.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
webrtcvad==2.0.10
